requirements.txt
django-tagging==0.3
-e git://github.com/jamesturk/django-markupfield.git#egg=django-markupfield
-e git://github.com/sunlightlabs/django-brainstorm.git#egg=django-brainstorm
-e git://github.com/sunlightlabs/django-feedinator.git#egg=django-feedinator
